requirements.txt
anyconfig==0.13.0
binpickle==0.3.4
blinker==1.6.2
cffi==1.15.1
click==8.1.3
csr==0.4.3
Flask==2.3.1
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
lenskit==0.14.2
llvmlite==0.39.1
MarkupSafe==2.1.2
msgpack==1.0.5
numba==0.56.4
numpy>=1.18,<1.24
pandas==1.5.3
psutil==5.9.5
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
scipy==1.10.1
seedbank==0.1.2
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.1
zipp==3.15.0
imdbpy==2022.7.9
Pillow==8.2.0
